model.url() can return a URL with an 'id'
I have found a way to fix my issue without modifying the adapter code. The only change you might want to consider is to remove the call to "model.url()". Currently, it can cause bugs because it can return a URL with an 'id' on the end. But the code in the adapter assumes there is no 'id' in the URL.